+++ Compose/samples/ComposeSamples/app/src/main/java/com/github/mag0716/composesamples/MainActivity.kt
@@ -3,36 +3,16 @@ package com.github.mag0716.composesamples
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.ScrollState
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.material.TopAppBar
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.compose.navigate
 import androidx.navigation.compose.rememberNavController
+import com.github.mag0716.composesamples.ui.Dummy
+import com.github.mag0716.composesamples.ui.SampleList
 import com.github.mag0716.composesamples.ui.theme.ComposeSamplesTheme
-import com.github.mag0716.composesamples.ui.theme.Typography
-
-// TODO: 実サンプルへの修正
-enum class Sample {
-    SAMPLE1,
-    SAMPLE2,
-    SAMPLE3,
-    SAMPLE4,
-    SAMPLE5,
-    SAMPLE6,
-    SAMPLE7,
-    SAMPLE8,
-    SAMPLE9,
-    SAMPLE10,
-}
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -40,6 +20,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             ComposeSamplesTheme {
                 Scaffold(
+                    // TODO: 画面遷移したらUp keyを表示させたいが、TopAppBarはどこで実装すべき？
                     topBar = {
                         TopAppBar(
                             title = { Text(getString(R.string.app_name)) }
@@ -51,43 +32,17 @@ class MainActivity : ComponentActivity() {
                         composable("sampleList") {
                             SampleList(
                                 Sample.values().toList()
-                            )
+                            ) {
+                                navController.navigate("dummy") {
+                                }
+                            }
+                        }
+                        composable("dummy") {
+                            Dummy()
                         }
                     }
                 }
             }
         }
     }
-}
-
-@Composable
-fun SampleList(samples: List<Sample>) {
-    Column(
-        modifier = Modifier.verticalScroll(ScrollState(0))
-    ) {
-        samples.forEach { sample ->
-            SampleItem(sample)
-            Divider()
-        }
-    }
-}
-
-@Composable
-fun SampleItem(sample: Sample) {
-    Text(
-        text = sample.name,
-        style = Typography.h4,
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable {
-                // TODO: 画面遷移処理
-            }
-            .padding(16.dp)
-    )
-}
-
-@Preview
-@Composable
-fun SampleItemPreview() {
-    SampleItem(Sample.SAMPLE1)
 }
