+++ Compose/samples/ComposeSamples/app/src/main/java/com/github/mag0716/composesamples/Sample.kt
@@ -1,0 +1,15 @@
+package com.github.mag0716.composesamples
+
+// TODO: 実サンプルへの修正
+enum class Sample {
+    SAMPLE1,
+    SAMPLE2,
+    SAMPLE3,
+    SAMPLE4,
+    SAMPLE5,
+    SAMPLE6,
+    SAMPLE7,
+    SAMPLE8,
+    SAMPLE9,
+    SAMPLE10,
+}
