+++ Compose/samples/ComposeSamples/app/src/main/java/com/github/mag0716/composesamples/ui/Dummy.kt
@@ -1,0 +1,10 @@
+package com.github.mag0716.composesamples.ui
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+// TODO: 実際のサンプルを実装したら削除する
+@Composable
+fun Dummy() {
+    Text("Dummy")
+}
